--- src/components/bookintroduction/newbook/NewBook.styled.ts
@@ -1,46 +1,0 @@
-import styled from 'styled-components';
-
-export const Section = styled.section`
-  display: flex;
-  justify-content: center;
-`;
-export const Container = styled.section`
-  width: 74%;
-`;
-export const Header = styled.header`
-  font-size: 64px;
-`;
-export const Body = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 20px;
-  padding: 20px;
-`;
-
-export const CategoryTitle = styled.div`
-  border-bottom: 2px solid #000;
-  font-size: 5rem;
-`;
-
-export const BookImageWrapper = styled.div``;
-export const BookGenre = styled.h3`
-  font-size: 14px;
-`;
-
-export const BookImg = styled.div``;
-export const BookIntro = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  justify-content: flex-end;
-`;
-export const Title = styled.h4``;
-export const AuthWrapper = styled.div`
-  display: flex;
-`;
-export const Author = styled.span``;
-export const Plot = styled.span``;
-
-export const BookWrapper = styled.div`
-  display: flex;
-`;
