+++ src/shared/Router.tsx
@@ -1,10 +1,6 @@
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 
-import BookBestseller from '../components/bookintroduction/bookbestseller/BookBestseller';
 import BookDetailPage from '../components/bookintroduction/bookdetailpage/BookDetailPage';
-import BookerPick from '../components/bookintroduction/bookerpick/BookerPick';
-import BookSpecial from '../components/bookintroduction/bookspecial/BookSpecial';
-import NewBook from '../components/bookintroduction/newbook/NewBook';
 
 import Layout from '../components/layout/Layout';
 
@@ -70,14 +66,15 @@ const Router = () => {
           {/* Qna 페이지 */}
           <Route path="/chat" element={<AdminChat />} />
           {/* 도서소개 페이지  */}
-          <Route element={<AboutBooks />}>
-            <Route path="/aboutbook/bestseller" element={<BookBestseller />} />
+          <Route path="/aboutbook/:id" element={<AboutBooks />} />
+          <Route path="/aboutbook/detail/:itemid" element={<BookDetailPage />} />
+
+          {/* <Route path="/aboutbook/bestseller" element={<BookBestseller />} />
             <Route path="/aboutbook/newbook" element={<NewBook />} />
             <Route path="/aboutbook/bookspecial" element={<BookSpecial />} />
             <Route path="/aboutbook/bookerpick" element={<BookerPick />} />
-          </Route>
+          </Route> */}
           {/* 도서 이동 페이지 */}
-          <Route path="/aboutbook/:itemid" element={<BookDetailPage />} />
         </Route>
 
         {/* 로그인/ 회원가입에 헤더 푸터 적용하고 싶으시면 Layout 라우터 태그 안에 넣어주시면 됩니다. */}
